% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_var.R
\name{barplot_var}
\alias{barplot_var}
\title{barplot_var}
\usage{
barplot_var(df, var_index)
}
\arguments{
\item{df}{Recibe de variable el  conjunto de datos de la data(package='palmerpenguins')}

\item{var_index}{Recibe una numeración para cada variable, definida en la funcion
1 = "species"
2 = "island"
3 = "sex"
"all" = Una grafica conjunta con las anteriores variables}
}
\value{
un grafico diagrama de barras de cada una de las variables de la data.
}
\description{
Esta función genera un diagrama de barras de la frecuencia de especies, islas y sexo de pingüinos.
Toma un marco de datos como argumento y devuelve un gráfico de barras.
}
\examples{
data(package='palmerpenguins')
df <- penguins
barplot_var(df, 1)
barplot_var(df, "all")
}
