% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_jitter_var.R
\name{boxplot_jitter_var}
\alias{boxplot_jitter_var}
\title{boxplot_jitter_var}
\usage{
boxplot_jitter_var(df, var_index)
}
\arguments{
\item{df}{Recibe de variable el  conjunto de datos de la data(package='palmerpenguins')}

\item{var_index}{Recibe una numeración para cada variable, definida en la funcion
1 = "bill_length_mm"
2 = "bill_depth_mm"
3 = "flipper_length_mm"
4 = "body_mass_g"
"all" = Una grafica conjunta con las anteriores variables}
}
\value{
un diagrama de dispersión
}
\description{
esta función genera un gráfico de dispersión para cada variable en el conjunto de datos de pingüinos,
con la opción de generar gráficas para variables individuales o para todas las variables.
}
\examples{
data(package='palmerpenguins')
df <- penguins
boxplot_jitter_var(df, 1)
boxplot_jitter_var(df, "all")

}
