% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plot.R
\name{correlation_plot}
\alias{correlation_plot}
\title{correlation_plot}
\usage{
correlation_plot(df)
}
\arguments{
\item{df}{Conjunto de datos que contiene las columnas ‘bill_length_mm’, ‘bill_depth_mm’, ‘flipper_length_mm’, y ‘body_mass_g’.}
}
\value{
una matriz de correlación
}
\description{
Esta función calcula la matriz de correlación de los atributos numéricos en un conjunto de datos
y traza un gráfico de correlación utilizando la función corrplot.
}
\examples{
data(package='palmerpenguins')
df <- penguins
correlation_plot(df)
}
