% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piechart_var.R
\name{piechart_var}
\alias{piechart_var}
\title{piechart_var}
\usage{
piechart_var(df, var_index)
}
\arguments{
\item{df}{Recibe de variable el  conjunto de datos de la data(package='palmerpenguins')}

\item{var_index}{Recibe una numeración para cada variable, definida en la funcion
1 = "species"
2 = "island"
3 = "sex"
"all" = Una grafica conjunta con las anteriores variables}
}
\value{
un gráfico circular de la frecuencia relativa
}
\description{
Esta función genera un gráfico circular de la frecuencia relativa de especies, islas y sexos de pingüinos.
Toma un marco de datos como argumento y devuelve un gráfico circular.
}
\examples{
data(package='palmerpenguins')
df <- penguins
piechart_var(df, 1)
piechart_var(df, "all")


}
